% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{clonetype}
\alias{clonetype}
\title{print clone type distribution for each niche}
\usage{
clonetype(nnt, file = "clonetype.pdf")
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}
}
\value{
pdf
}
\description{
print clone type distribution for each niche
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
clonetype(nnt)
}
