% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{print_nichenetwork}
\alias{print_nichenetwork}
\title{print clone network}
\usage{
print_nichenetwork(
  nnt,
  file = "nichenetwork.pdf",
  file_layout = "layout.csv",
  ilayout = NULL,
  highlight = FALSE,
  seed = 888,
  mark_groups = FALSE,
  margin = c(0, 0, 0, 0),
  vertex.label.cex = 1,
  vsize = 1,
  esize = 1,
  ecolor = "grey80",
  width = 7,
  height = 7,
  weighted = FALSE,
  direction = FALSE,
  niter = 1e+05,
  axes = FALSE
)
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf file name to be saved}

\item{file_layout}{file name to save layout of clones}

\item{ilayout}{file used to set external layout}

\item{highlight}{whether to hightlight clones for each tag in pdf file}

\item{seed}{seed for random clone colour setting}

\item{mark_groups}{igrap parameter}

\item{margin}{igraph parameter}

\item{vertex.label.cex}{vertex label size}

\item{vsize}{vertex size}

\item{esize}{edge size}

\item{ecolor}{edge color}

\item{width}{file size}

\item{height}{file size}

\item{weighted}{layout parameter}

\item{direction}{layout parameter}

\item{niter}{layout parameter}

\item{axes}{TorF, if or not draw axes}
}
\value{
df.niche2cloneID
}
\description{
print clone network
}
\examples{
data(example)
nnt=Dnichenetwork(scObject,groupby="cell_clusters")
print_nichenetwork(nnt,file="nichenetwork-w3-26.pdf",vertex.label.cex=0.1,vsize=1.3,esize=1,ecolor=adjustcolor("grey80", alpha.f = 0.5),mark_groups = TRUE)
print_nichenetwork(nnt,file="nichenetwork-w3-27.pdf",vertex.label.cex=0.1,vsize=1.3,esize=1,ecolor=NA,mark_groups = TRUE)
print_nichenetwork(nnt,file="nichenetwork-w3-28.pdf",vertex.label.cex=0.1,vsize=1.3,esize=1,ecolor=NA)
print_nichenetwork(nnt,file="nichenetwork-w3-29.pdf",vertex.label.cex=0.1,vsize=1.3,esize=1)
}
