% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_niche.R
\name{tag_cellclonetype}
\alias{tag_cellclonetype}
\title{print cell type and clone type distribution for each niche}
\usage{
tag_cellclonetype(nnt, file = "tag2celltype_clonetype.pdf", seed = 888)
}
\arguments{
\item{nnt}{a list contains connectome information of all niches, the result of Dnichenetwork}

\item{file}{pdf}
}
\value{
pdf
}
\description{
print cell type and clone type distribution for each niche
}
\examples{
tag_cellclonetype(nnt)
}
